+++ pkg/k8s/storage.go
@@ -2,90 +2,76 @@ package k8s
 
 import (
 	"context"
+	"errors"
 
 	"github.com/blang/semver/v4"
+	v1 "k8s.io/api/storage/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-type StorageProvision struct {
-	PersistentStorage bool
-	Reason            string
-}
+const (
+	AWS_EBS_CSI_DRIVER_NAME               = "ebs.csi.aws.com"
+	AWS_EBS_STORAGE_CLASS_NOT_DEFAULT     = "found default storage class without aws-ebs provisioner"
+	DEFAULT_STORAGE_CLASS_ANNOTATION_NAME = "storageclass.kubernetes.io/is-default-class"
 
-func GenerateStorageProvision(ctx context.Context, client *Client, clusterSummary *ClusterSummary) StorageProvision {
+	HINT_INSTALL_AWS_EBS_CSI_DRIVER = `Hint: 
+  * Install Amazon EBS CSI driver: https://docs.aws.amazon.com/eks/latest/userguide/ebs-csi.html`
+	HINT_DEFINE_DEFAULT_STORAGE_CLASS = `Hint:
+  * Define default StorageClass: https://kubernetes.io/docs/concepts/storage/storage-classes/#the-storageclass-resource`
+)
+
+var (
+	ErrNoDefaultStorageClass = errors.New("cluster has no default storage class")
+)
+
+func (clusterRequirements ClusterRequirements) validateStorage(ctx context.Context, client *Client, clusterSummary *ClusterSummary) Requirement {
+	var err error
+
+	var requirement Requirement
+	requirement.Message = CLUSTER_STORAGE_SUPPORTED
+
+	if _, err = getDefaultStorageClass(ctx, client); err != nil {
+		requirement.IsCompatible = false
+		requirement.IsNonCompatible = true
+		requirement.ErrorMessages = append(requirement.ErrorMessages, err.Error(), HINT_DEFINE_DEFAULT_STORAGE_CLASS)
+		return requirement
+	}
+
 	if IsEksCluster(clusterSummary.ClusterName) {
-		return generateEksStorageProvision(ctx, client, clusterSummary)
+		if semver.MustParseRange("<1.23.0")(clusterSummary.ServerVersion) {
+			requirement.IsCompatible = len(requirement.ErrorMessages) == 0
+			return requirement
+		}
+
+		if err = hasEbsCsiDriver(ctx, client); err != nil {
+			requirement.IsCompatible = false
+			requirement.IsNonCompatible = true
+			requirement.ErrorMessages = append(requirement.ErrorMessages, err.Error(), HINT_INSTALL_AWS_EBS_CSI_DRIVER)
+			return requirement
+		}
 	}
 
-	return generateDefaultStorageProvision(ctx, client, clusterSummary)
+	requirement.IsCompatible = len(requirement.ErrorMessages) == 0
+
+	return requirement
 }
 
-func hasDefaultStorageClass(ctx context.Context, client *Client, clusterSummary *ClusterSummary) bool {
-	storageClasses, err := client.StorageV1().StorageClasses().List(ctx, metav1.ListOptions{})
+func getDefaultStorageClass(ctx context.Context, client *Client) (*v1.StorageClass, error) {
+	storageClassList, err := client.StorageV1().StorageClasses().List(ctx, metav1.ListOptions{})
 	if err != nil {
-		return false
+		return nil, err
 	}
 
-	for _, storageClass := range storageClasses.Items {
-		if storageClass.Annotations["storageclass.kubernetes.io/is-default-class"] == "true" {
-			return true
+	for _, storageClass := range storageClassList.Items {
+		if value, ok := storageClass.Annotations[DEFAULT_STORAGE_CLASS_ANNOTATION_NAME]; ok && value == "true" {
+			return &storageClass, nil
 		}
 	}
 
-	return false
+	return nil, ErrNoDefaultStorageClass
 }
 
-func generateEksStorageProvision(ctx context.Context, client *Client, clusterSummary *ClusterSummary) StorageProvision {
-	if clusterSummary.ServerVersion.LT(semver.MustParse("1.23.0")) {
-		return StorageProvision{
-			PersistentStorage: true,
-			Reason:            "Kubernetes version is less than 1.23.0",
-		}
-	}
-
-	list, err := client.CoreV1().Pods(metav1.NamespaceAll).List(ctx, metav1.ListOptions{
-		LabelSelector: "app.kubernetes.io/name=aws-ebs-csi-driver",
-	})
-
-	if err != nil {
-		return StorageProvision{
-			PersistentStorage: false,
-			Reason:            "Error listing aws csi driver pods",
-		}
-	}
-
-	if len(list.Items) == 0 {
-		return StorageProvision{
-			PersistentStorage: false,
-			Reason:            "No aws csi driver pods found",
-		}
-	}
-
-	hasStorageClass := hasDefaultStorageClass(ctx, client, clusterSummary)
-	if hasStorageClass {
-		return StorageProvision{
-			PersistentStorage: true,
-			Reason:            "Has default storage class",
-		}
-	}
-
-	return StorageProvision{
-		PersistentStorage: false,
-		Reason:            "Has aws ebs dirver without default storage class",
-	}
-
-}
-
-func generateDefaultStorageProvision(ctx context.Context, client *Client, clusterSummary *ClusterSummary) StorageProvision {
-	if hasDefaultStorageClass(ctx, client, clusterSummary) {
-		return StorageProvision{
-			PersistentStorage: true,
-			Reason:            "Has default storage class",
-		}
-	}
-
-	return StorageProvision{
-		PersistentStorage: false,
-		Reason:            "Cluster has no default storage class",
-	}
+func hasEbsCsiDriver(ctx context.Context, client *Client) error {
+	_, err := client.StorageV1().CSIDrivers().Get(ctx, AWS_EBS_CSI_DRIVER_NAME, metav1.GetOptions{})
+	return err
 }